[fix] refactor and add comments
